changed from template to active page
took out the placeholder for the image replaced it with an active image, and changed all the garbage text for an actual text from a burger shop called Burger Smash! Added icons instead of empty boxes for the socials page, need to change the socials page to be about us or what to expect. changed the quote to a quote from the founder of Burger Smash! and changed the copyright on the footer of the page to the proper naming